create blog article template
